+++ CHANGELOG.md
@@ -21,13 +21,13 @@ for further information about branching and tagging conventions.
   - New `KycRequestRecord`
   - New `KycRecoveryRequestRecord`
   - New `RequestRecord` getters: `isExists`, `updatableId`
+- Formers:
+  - New `CreateAssetFormer`
 
 #### Changed
 - Extract `initApi()`, call it from the main.js, instead of App.vue
   (the api should be initialized before the App rendered)
 - `loadingDataViaLoop()` => `loadAllResponsePages()`, moved to `api-helpers.js`
-- Formers:
-  - New `KycGeneralFormer` and `KycCorporateFormer`
 - Proxied `DOCUMENT_TYPES` via Vue prototype as `$DOCUMENT_TYPES`
 - `vuexTypes.LOAD_ACCOUNT`’s `accountId` arg is optional now
 
@@ -36,6 +36,8 @@ for further information about branching and tagging conventions.
 - Old key-value module
 - Old `ChangeRoleRecord`
 - Old `verification-form.mixin.js`
+- Modules:
+  - `create-asset-form-simplified`
 
 #### Fixed
 - A bug with `MarkdownField` length